api/requirements.txt
flask>=3.0
flask-cors>=4.0
openai-whisper
